django/requirements.txt
Django==4.2.0
django-cors-headers>=4.2.0
djangorestframework>=3.11,<4.0
psycopg2-binary>=2.8
redis>=4.0
requests>=2.31.0
celery>=5.1.2
drf-yasg>=1.21.0
channels>=4.0.0
channels-redis>=4.1
uvicorn>=0.20.0
websockets>=9.1
gunicorn>=21.1.0
